Revert deleting Shippable Purchase Item behavior.
It is used for shippable product purchase items and
not for shipping purchase items.

The difference is that it would be better to attach the behavior only to
purchase items which need the shipping_item association and not all of them.

In the tests we are applying it to Model_Purchase_Item_Product.