Remove dependency on freezable behavior
Until now the freezable behavior configuration was modified on the fly
to tell store purchase models which used the shippable store purchase behavior
to freeze the shipping association as well.

Users of the shipping module would need to do this by themselves usign the freezable traits
and their abstract methods.